Move Juno.java and Messages.java to juno/main/
